refactor out skeleton
